Break up client/descriptions.go
